example/requirements.txt
Django==4.1.6
djangorestframework==3.14.0
drf-simple-apikey==0.1.2
